move search functionality to a separate file
