+++ internal/k8s/k8s.go
@@ -11,15 +11,11 @@ import (
 	"github.com/google/uuid"
 	"github.com/nais/api/internal/auth/authz"
 	"github.com/nais/api/internal/database"
-	"github.com/nais/api/internal/graph/model"
-	"github.com/nais/api/internal/search"
 	"github.com/nais/api/internal/slug"
 	kafka_nais_io_v1 "github.com/nais/liberator/pkg/apis/kafka.nais.io/v1"
 	naisv1 "github.com/nais/liberator/pkg/apis/nais.io/v1"
 	naisv1alpha1 "github.com/nais/liberator/pkg/apis/nais.io/v1alpha1"
 	"github.com/sirupsen/logrus"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/dynamic/dynamicinformer"
@@ -239,111 +235,6 @@ func New(tenant string, cfg Config, db Database, log logrus.FieldLogger, opts ..
 	}, nil
 }
 
-func (c *Client) Search(ctx context.Context, q string, filter *model.SearchFilter) []*search.Result {
-	if !isFilterOrNoFilter(filter) {
-		return nil
-	}
-
-	if c.database == nil {
-		panic("database not set")
-	}
-
-	ret := []*search.Result{}
-
-	for env, infs := range c.informers {
-		if isFilterSqlInstanceOrNoFilter(filter) {
-			if infs.SqlInstanceInformer == nil {
-				continue
-			}
-
-			sqlInstances, err := infs.SqlInstanceInformer.Lister().List(labels.Everything())
-			if err != nil {
-				c.error(ctx, err, "listing SQL instances")
-				return nil
-			}
-
-			for _, obj := range sqlInstances {
-				u := obj.(*unstructured.Unstructured)
-				rank := search.Match(q, u.GetName())
-				if rank == -1 {
-					continue
-				}
-
-				sqlInstance, err := model.ToSqlInstance(u, env)
-				if err != nil {
-					c.error(ctx, err, "converting to SQL instance")
-					return nil
-				} else if ok, _ := c.database.TeamExists(ctx, sqlInstance.GQLVars.TeamSlug); !ok {
-					continue
-				}
-
-				ret = append(ret, &search.Result{
-					Node: sqlInstance,
-					Rank: rank,
-				})
-			}
-		}
-
-		if isFilterNaisjobOrNoFilter(filter) {
-			jobs, err := infs.NaisjobInformer.Lister().List(labels.Everything())
-			if err != nil {
-				c.error(ctx, err, "listing jobs")
-				return nil
-			}
-
-			for _, obj := range jobs {
-				u := obj.(*unstructured.Unstructured)
-				rank := search.Match(q, u.GetName())
-				if rank == -1 {
-					continue
-				}
-				job, err := c.ToNaisJob(u, env)
-				if err != nil {
-					c.error(ctx, err, "converting to job")
-					return nil
-				} else if ok, _ := c.database.TeamExists(ctx, job.GQLVars.Team); !ok {
-					continue
-				}
-
-				ret = append(ret, &search.Result{
-					Node: job,
-					Rank: rank,
-				})
-			}
-		}
-
-		if isFilterAppOrNoFilter(filter) {
-			apps, err := infs.AppInformer.Lister().List(labels.Everything())
-			if err != nil {
-				c.error(ctx, err, "listing applications")
-				return nil
-			}
-
-			for _, obj := range apps {
-				u := obj.(*unstructured.Unstructured)
-				rank := search.Match(q, u.GetName())
-				if rank == -1 {
-					continue
-				}
-				app, err := c.toApp(ctx, u, env)
-				if err != nil {
-					c.error(ctx, err, "converting to app")
-					return nil
-				} else if ok, _ := c.database.TeamExists(ctx, app.GQLVars.Team); !ok {
-					continue
-				}
-
-				ret = append(ret, &search.Result{
-					Node: app,
-					Rank: rank,
-				})
-			}
-		}
-
-	}
-	return ret
-}
-
 // Informers returns a map of informers, keyed by environment
 func (c *Client) Informers() ClusterInformers {
 	return c.informers
@@ -365,48 +256,4 @@ func convert(m any, target any) error {
 func (c *Client) error(_ context.Context, err error, msg string) error {
 	c.log.WithError(err).Error(msg)
 	return fmt.Errorf("%s: %w", msg, err)
-}
-
-func isFilter(filter *model.SearchFilter) bool {
-	if filter == nil {
-		return false
-	}
-
-	if filter.Type == nil {
-		return false
-	}
-
-	return true
-}
-
-func isFilterOrNoFilter(filter *model.SearchFilter) bool {
-	if !isFilter(filter) {
-		return true
-	}
-
-	return *filter.Type == model.SearchTypeApp || *filter.Type == model.SearchTypeNaisjob || *filter.Type == model.SearchTypeSQLInstance
-}
-
-func isFilterAppOrNoFilter(filter *model.SearchFilter) bool {
-	if !isFilter(filter) {
-		return true
-	}
-
-	return *filter.Type == model.SearchTypeApp
-}
-
-func isFilterNaisjobOrNoFilter(filter *model.SearchFilter) bool {
-	if !isFilter(filter) {
-		return true
-	}
-
-	return *filter.Type == model.SearchTypeNaisjob
-}
-
-func isFilterSqlInstanceOrNoFilter(filter *model.SearchFilter) bool {
-	if !isFilter(filter) {
-		return true
-	}
-
-	return *filter.Type == model.SearchTypeSQLInstance
 }
